% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DynamicPortfolioManagement.R
\name{BellmanEq_CS1}
\alias{BellmanEq_CS1}
\title{Solves the Bellman Equation for the case study 1.}
\usage{
BellmanEq_CS1(eta, gamma, lambda, tau, theta, mu, sig2, c2, b_legacy, x, t_view,
  view)
}
\arguments{
\item{eta}{[scalar] overall weight of the market impact of transactions}

\item{gamma}{[scalar] risk aversion parameter}

\item{lambda}{[scalar] discounting parameter}

\item{tau}{[scalar] trading interval}

\item{theta}{[n_ x n_] transition matrix of the MVOU process}

\item{mu}{[n_ x 1]  drift vector of the MVOU process}

\item{sig2}{[n_ x n_] covariance parameters of the MVOU process}

\item{c2}{[n_ x n_] matrix of the market impact}

\item{b_legacy}{[n_ x 1]  legacy portfolio exposure at time 0}

\item{x}{[t_ x n_] path of the risk drivers (with time step = tau)}

\item{t_view}{[1 x N_MeanViews] times of the views}

\item{view}{[1 x N_MeanViews] views on the risk drivers}
}
\value{
prior  [t_x n_ matrix] optimal prior exposure

post   [t_x n_ matrix] optimal posterior exposure
}
\description{
Solves the Bellman Equation for the case study 1.
}
\details{
In Case Study 1 there is only one risk driver (the 10 year rate) and
only one view. The view is that the expected value of the 10-year rate will
be the actual value minus 50 basis points at t^view = 1 year from the current
time.
The solution is analytical in the prior case.
The solution is found recursivelly in the posterior case starting from
t_view and going back to time 0 with a time step = tau.

In case study 1: n_ = 1 N_meanViews = 1
}
\author{
Xavier Valls \email{xavievallspla@gmail.com}
}
\references{
A. Meucci - "Dynamic Portfolio Management with Views at Multiple Horizons"
\url{http://symmys.com/node/831}. See Meucci script for "BellmanEq_CS1.m"
}

