\docType{data}
\name{bondAttribution}
\alias{bondAttribution}
\title{bondAttribution}
\description{
  bondAttribution
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

