% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterfliesAnalytics}
\alias{butterfliesAnalytics}
\title{list of securities with analytics computed.}
\description{
list of securities with analytics computed.
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, "Fully Flexible Views: Theory and Practice", The Risk Magazine,
October 2008, p 100-106. \url{http://symmys.com/node/158}
}
\keyword{data}

