% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAsset.R
\name{createAsset}
\alias{createAsset}
\title{Create a mAsset (Meucci Asset) object containing time series and asset-specific meta-data}
\usage{
createAsset(assetName = NULL, assetTimeSeries = NULL,
  assetTimeSeriesType = c("level", "difference", "discrete", "log"),
  assetClass = c("equity", "debt", "currency", "option", "future"),
  pricingFunction = NULL, seedValue = NULL, invariant = list(type =
  c("log", "difference", "residual"), model = list(data = NULL, model = NULL)),
  invariantFunction = quote(Return.calculate.mAsset),
  marginalDistribution = NULL)
}
\arguments{
\item{assetName}{a string naming the object}

\item{assetTimeSeries}{an xts object of the raw series; prices/levels are preferred, but returns can be passed}

\item{assetTimeSeriesType}{one of "level", "difference", "discrete", "log", so that the package knows how to handle the xts object}

\item{assetClass}{one of "equity", "debt", "currency", "option", "future"}

\item{pricingFunction}{a function, currently not used}

\item{seedValue}{the price or level associated with the first time-stamp in the assetTimeSeries object, if levels/prices are provided it will automatically be extracted}

\item{invariant}{a list of invariant information - this is experimental and may likely change as development progresses}

\item{invariantFunction}{a function used to calculate the assetTimeSeries argument to the asset-specific invariant - again this is experimental and may likely change as development progresses}

\item{marginalDistribution}{a fitted distribution function - not yet used, may move to "invariant"}
}
\description{
This function effectively adds extensive metadata to xts objects, linking asset-specific details to the object itself, simplifying 
Meucci-method calculations by using data defined at object-creation.  It requires a name, time series, 
time series type, asset class, and invariant information.  By providing this information, 
executing functions is simplified, since each function is executed according to Meucci's methods.
The mAsset objects are the building blocks for mPortfolio objects.
}
\seealso{
\code{\link{createPortfolio}}
}
\author{
Erol Biceroglu
}
