% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPortfolio.R
\name{createPortfolio}
\alias{createPortfolio}
\title{Create a mPortfolio (Meucci Portfolio) object linking mAsset objects into one}
\usage{
createPortfolio(..., join = "fullouter", multivariateDistribution = NULL)
}
\arguments{
\item{...}{mAsset objects}

\item{join}{defaults to "fullouter"}

\item{multivariateDistribution}{an optional multivariate distribution for the invariants - this may be removed as development progresses}
}
\description{
This function combines mAsset objects into an mPortfolio, and uses additional user-defined portfolio-level metadata,
thereby simplifying Meucci-method calculations by using data defined at object-creation.
It requires at least 1 mAsset object, defaults to an outer join, and potentially a multivariate distribution for the invariants.
Note for advanced users, each mAsset object is stored on its own within a list, and so the xts matrix is constructed on-demand.
 
Product of all the individual period returns
}
\details{
For arithmetic returns:
\deqn{(1+r_{1})(1+r_{2})(1+r_{3})\ldots(1+r_{n})=cumprod(1+R)}{cumprod(1+R)}

For log returns:
\deqn{exp(r_{1}+r_{2}+r_{3} + \ldots + r_{n})=exp(cumsum(R))}{exp(cumsum(R))}
}
\examples{

#' TBD... this isn't related
data(managers)
managers <- managers[,1:6] #equities only
xtsAttributes(managers) <- list(coredata_content = "discreteReturn") #append attribute
chart.CumReturns(managers) #here is baseline
managersL <- Level.calculate(R = managers)
plot(managersL-1)

#Here they are equal
Return.cumulative(managers)  #Baseline
last(managersL-1) #This function

}
\seealso{
\code{\link{createAsset}}
}
\author{
Erol Biceroglu
}
