% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbFFP}
\alias{dbFFP}
\title{Historical Scenarios with Fully Flexible Probabilities dataset.}
\description{
Data for the Historical Scenarios with Fully Flexible Probabilities paper.
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, "Personalized Risk Management: Historical Scenarios with Fully Flexible Probabilities"
GARP Risk Professional, Dec 2010, p 47-51. \url{http://www.symmys.com/node/150}
}
\keyword{data}

