% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dynamicManagement}
\alias{dynamicManagement}
\title{data for the dynamic portfolio management examples}
\description{
data for the dynamic portfolio management examples
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, D. Ardia, S. Keel - "Fully Flexible Extreme Views"
}
\keyword{data}

