% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EffectiveBets.R
\name{EffectiveBets}
\alias{EffectiveBets}
\title{computes the Effective Number of Bets and the Diversification
distribution}
\usage{
EffectiveBets(b, Sigma, t)
}
\arguments{
\item{b}{[vector] (n_ x 1) exposures}

\item{Sigma}{[matrix] (n_ x n_) covariance matrix}

\item{t}{[matrix] (n_ x n_) torsion matrix}
}
\value{
enb       [scalar] Effetive Number of Bets

p         [vector] (n_ x 1) diversification distribution
}
\description{
computes the Effective Number of Bets and the Diversification
distribution, as described in A. Meucci, A. Santangelo, R. Deguest -
"Measuring Portfolio Diversification Based on Optimized Uncorrelated Factors"
}
\author{
Xavier Valls \email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, A. Santangelo, R. Deguest - "Measuring Portfolio Diversification
Based on Optimized Uncorrelated Factors" \url{http://symmys.com/node/599}

See Meucci's script "EffectiveBets.m"
}

