\docType{data}
\name{Equities}
\alias{Equities}
\title{daily stock prices from the utility sector in the S&P 500}
\description{
  daily stock prices from the utility sector in the S&P
  500.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

