\docType{data}
\name{FDButterflies}
\alias{FDButterflies}
\title{Factor Distribution Butterflies}
\description{
  Factor Distribution Butterflies
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, "Fully Flexible Views: Theory and Practice",
  The Risk Magazine, October 2008, p 100-106.
  \url{http://symmys.com/node/158}
}
\keyword{data}
\keyword{datasets}

