\docType{data}
\name{fILMR}
\alias{fILMR}
\title{Fully Integrated Liquidity and Market Risk Model dataset.}
\description{
  Data for the Fully Integrated Liquidity and Market Risk
  Model paper.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, "A Fully Integrated Liquidity and Market Risk
  Model", Financial Analyst Journal, 68, 6, 35-47 (2012)
  \url{http://www.symmys.com/node/350}
}
\keyword{data}
\keyword{datasets}

