\docType{data}
\name{fixedIncome}
\alias{fixedIncome}
\title{US government yield curve and bond yield data}
\description{
  US government yield curve and bond yield data for
  different dates.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

