\docType{data}
\name{freaqEst}
\alias{freaqEst}
\title{scenarios table and prior distribution for changes in SWAP2YR SWAP10YR CDXIG S&P500 DollarIndex Crude Gold VIX 10YRInflationSwapRate}
\description{
  scenarios table and prior distribution for changes in
  SWAP2YR SWAP10YR CDXIG S&P500 DollarIndex Crude Gold VIX
  10YRInflationSwapRate
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, "Stress-Testing with Fully Flexible Causal
  Inputs" \url{http://symmys.com/node/152}
}
\keyword{data}
\keyword{datasets}

