\name{garch1f4}
\alias{garch1f4}
\title{Fit a GARCH(1,1) model with student-t errors}
\usage{
  garch1f4(x, eps, df)
}
\arguments{
  \item{x}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}

  \item{eps}{: [scalar] used in enforcing a_ii + b_ii <= 1
  - eps; the default value is zero}

  \item{df}{: [scalar] degree of freedom for the
  t-distribution; the default value is 500 to make it,
  basically, normal}
}
\value{
  q : [vector] (4 x 1) parameters of the GARCH(1,1) process

  qerr : [vector] (4 x 1) standard error of parameter
  estimates

  hf : [scalar] current conditional heteroskedasticity
  estimate

  hferr : [scalar] standard error on hf
}
\description{
  Fit a GARCH(1,1) model with student-t errors
}
\note{
  MATLAB's script initially written by Olivier Ledoit,
  4/28/1997 Uses a conditional t-distribution with fixed
  degrees of freedom Difference with garch1f: errors come
  from the score alone
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  \url{http://symmys.com/node/170} See Meucci's script for
  "FitMultivariateGarch.m"
}

