\name{garch2f8}
\alias{garch2f8}
\title{Off-diagonal parameter estimation in bivariate GARCH(1,1) when diagonal parameters are given.}
\usage{
  garch2f8(y, c1, a1, b1, y1, h1, c2, a2, b2, y2, h2, df)
}
\arguments{
  \item{x}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}
}
\value{
  q : [vector] (4 x 1) parameters of the GARCH(1,1) process

  qerr : [vector] (4 x 1) standard error of parameter
  estimates

  hf : [scalar] current conditional heteroskedasticity
  estimate

  hferr : [scalar] standard error on hf
}
\description{
  Off-diagonal parameter estimation in bivariate GARCH(1,1)
  when diagonal parameters are given.
}
\note{
  Initially written by Olivier Ledoit, 4/28/1997 Uses a
  conditional t-distribution with fixed degrees of freedom
  Steepest Ascent on boundary, Hessian off boundary, no
  grid search
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  \url{http://symmys.com/node/170} See Meucci's script for
  "FitMultivariateGarch.m"
}

