\name{GenFirstEigVect}
\alias{GenFirstEigVect}
\title{This function generates the first eigen vector}
\usage{
  GenFirstEigVect(S, A)
}
\arguments{
  \item{S}{Covariance Matrix}

  \item{A}{Conditioning Matrix}
}
\value{
  e First Eigen Vector
}
\description{
  This function generates the first eigen vector
}
\author{
  Manan Shah \email{mkshah@cmu.edu}
}
\references{
  A. Meucci - "Managing Diversification", Risk Magazine,
  June 2009 \url{http://ssrn.com/abstract=1358533}
}

