% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghqx}
\alias{ghqx}
\title{GHQX: Gauss-Hermite grid demo's input data}
\description{
GHQX: Gauss-Hermite grid demo's input data
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, D. Ardia, S. Keel - "Fully Flexible Extreme Views"
}
\keyword{data}

