\name{integrateSubIntervals}
\alias{integrateSubIntervals}
\title{Integrate the subinterval for the given cumulative distribution function to get the equivalent probability}
\usage{
  integrateSubIntervals(x, cdf)
}
\arguments{
  \item{x}{a vector containing the data points}

  \item{cdf}{the cumulative distribution function}
}
\value{
  p a vector containing the cdf evaluated for each of the
  subintervals
}
\description{
  Integrate the subinterval for the given cumulative
  distribution function to get the equivalent probability
}
\author{
  Ram Ahluwalia \email{ram@wingedfootcapital.com}
}

