% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InverseCallTransformation.R
\name{InverseCallTransformation}
\alias{InverseCallTransformation}
\title{Computes the Inverse Call Transformation}
\usage{
InverseCallTransformation(rates, tau, eta, zeta)
}
\arguments{
\item{Rates}{[matrix] (N x length(timeseries)) number of portfolio
in the efficient frontier}

\item{Tau}{[vector] (N x 1) vector containing the times to
maturity corresponding to the rows of the rates matrix}

\item{Eta}{[scalar] Inverse-call transformation parameter}

\item{Zeta}{[scalar] Inverse-call transformation parameter}
}
\value{
x               [vector] (NumPortf x 1) shadow rates, computed from
rates via inverse-call transformation
}
\description{
Computes the Inverse Call Transformation and returns shadow
rates, as described in A. Meucci, A. Loregian -"Neither Normal not Lognormal:
Modeling Interest Rates Across all Regimes".
}
\note{
Smoothing parameter s obtained as s=eta*exp(zeta*tau)
}
\author{
Xavier Valls \email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, A. Loregian - "Neither Normal not Lognormal: Modeling Interest
Rates Across all Regimes" \url{http://symmys.com/node/601}

See Meucci's script "InverseCallTransformation.m".
}

