% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JGB}
\alias{JGB}
\title{japanese government bonds}
\description{
japanese government bonds
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, A. Loregian - "Neither Normal not Lognormal: Modeling Interest Rates Across all Regimes" \url{http://symmys.com/node/601}
}
\keyword{data}

