\name{kernelbw}
\alias{kernelbw}
\title{Generates bandwidth of a Gaussian Kernel Density Estimator based on Silverman's rule of thumb}
\usage{
  kernelbw(xi)
}
\arguments{
  \item{xi}{a vector containing the data set for which the
  bandwidth has to be calculated}
}
\value{
  bw a numeric signifying the bandwidth
}
\description{
  Generates bandwidth of a Gaussian Kernel Density
  Estimator based on Silverman's rule of thumb
}
\author{
  Manan Shah \email{mkshah@tepper.cmu.edu}
}

