% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linRet}
\alias{linRet}
\title{equal-weight portfolio of stocks in the S&P500}
\description{
equal-weight portfolio of stocks in the S&P500
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, A. Santangelo, R. Deguest - "Measuring Portfolio Diversification Based on Optimized Uncorrelated Factors" \url{http://symmys.com/node/599}
}
\keyword{data}

