% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MVOUPosterior.R
\name{MVOU_Posterior}
\alias{MVOU_Posterior}
\title{Computes the posterior conditional expectation & covariance matrix of}
\usage{
MVOU_Posterior(Mom, Views)
}
\arguments{
\item{Mom}{list  list of risk drivesrs}

\item{Views}{list  list of views}
}
\value{
Posterior    list of posterior distribution information
}
\description{
This function computes the  posterior conditional expectation
and covariance matrix at different monitoring dates t1,t2,...,t_ of the process
X_{t1,t2...,t_} =(X_t1,
                  X_t2,
                  .
                  .
                  X_t_)
X_t follows a MVOU process: dX_t <- (-theta*X_t+mu)dt + sig*dB_t
The Views are: E{v_mu*X}=mu_view and Cov{v_sig*X}=sig2_view
}
\details{
The list Mom consists of.
\itemize{
\item{Mom$monitoring_time }{ monitoring times =t_*n_ x 1]}
\item{Mom$dimension }{ labels of the risk drivers [t_*n_ x 1]}
\item{Mom$cov }{ prior covariance matrix of X_\{t1,t2...t_\} [t_*n_ x t_*n_]}
\item{Mom$mean }{ prior vector of the means of X_\{t1,t2...t_\} [t_*n_ x 1]}
\item{Mom$mean_cost}{[t_*n_ x 1]}
\item{Mom$mean_lin}{[t_*n_ x 1]}
\item{}{Mom$mean_cost and Mom$mean_lin are such that Mom$mean_cost +
Mom$mean_lin*x0  = Mom$mean}}

while the list of Views has the elements:
\itemize{
\item{Views$N_MeanViews }{ Number of views on the expectations [scalar]}
\item{Views$N_CovViews }{ Number of views on the covariance matrix [scalar]}
\item{Views$dimension }{ labels of the risk drivers [t_*n_ x 1]}
\item{Views$monitoring_time }{ monitoring times [t_*n_ x 1]}
\item{Views$v_mu }{ matrix that qualifies the views on expectations
             [N_MeanViews x t_*n_]}
\item{Views$v_sig }{ matrix that qualifies the views on covariance
             [N_CovViews x t_*n_]}
\item{Views$mu_view }{ extent of the views on expectation [N_MeanViews x 1]}
\item{Views$sig2_view }{ extent of the views on the covariance
 [N_CovViews x 1]}}

And the returned Posterior distribution list includes the elements:
\itemize{
\item{Posterior$monitoring_time }{ monitoring times [t_*n_ x 1]}
\item{Posterior$dimension }{ labels of the risk drivers [t_*n_ x 1]}
\item{Posterior$cov }{ posterior covariance matrice of X_{t1,t2...t_}
[t_*n_ x t_*n_]}
\item{Posterior$mean }{ posterior vector of the means of X_{t1,t2...t_}
[t_*n_ x 1]}
\item{Posterior$mean_cost [t_*n_ x 1]}
\item{Posterior$mean_lin [t_*n_ x 1]}}
Posterior$mean_cost and Posterior$mean_lin are such that
Posterior$mean_cost + Posterior$mean_lin*x0  = Posterior$mean
}
\author{
Xavier Valls \email{xavievallspla@gmail.com}
}
\references{
A. Meucci - "Dynamic Portfolio Management with Views at Multiple Horizons"
\url{http://symmys.com/node/831}. See Meucci script for "MVOU_Prior.m"
}

