% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MVOUPrior.R
\name{MVOU_Prior}
\alias{MVOU_Prior}
\title{Computes the conditional mean and covariance matrix at different
dates}
\usage{
MVOU_Prior(t, x0, theta, sig2, mu)
}
\arguments{
\item{t}{[t_ x 1]  vector of monitoring dates}

\item{x0}{[n_ x 1]  observation at time 0}

\item{theta}{[n_ x n_] transition matrix}

\item{sig2}{[n_ x n_] covariance matrix}

\item{mu}{[n_ x 1]  vector of drift parameters}
}
\value{
Mom     list of the risk drivers
}
\description{
This function computes the conditional mean and covariance
matrix at different monitoring dates t1,t2,...,t_ of  the process
X_{t1,t2...,t_} =(X_t1,
                  X_t2,
                  .
                  .
                  X_t_)
X_t follows a MVOU process: dX_t <- (-theta*X_t+mu)dt + sig*dB_t
}
\author{
Xavier Valls \email{xavievallspla@gmail.com}
}
\references{
A. Meucci - "Dynamic Portfolio Management with Views at Multiple Horizons"
\url{http://symmys.com/node/831}. See Meucci script for "MVOU_Prior.m"
}

