\name{normalizeProb}
\alias{normalizeProb}
\title{Generates the normalized probability for an input probability value}
\usage{
  normalizeProb(p)
}
\arguments{
  \item{p}{a numeric value containing the probability value
  required to be normalized}
}
\value{
  normalizedp a numeric value containing the normalized
  probability value
}
\description{
  Generates the normalized probability for an input
  probability value
}
\author{
  Ram Ahluwalia \email{ram@wingedfootcapital.com}
}

