% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EntropyProg.R
\name{PHist}
\alias{PHist}
\title{Generates histogram}
\usage{
PHist(X, p, nBins, freq = FALSE, main = "Portfolio return distribution",
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{X}{vector containing the data points}

\item{p}{vector containing the probabilities for each of the data
points in X}

\item{nBins}{expected number of Bins the data set is to be broken down into}

\item{freq}{boolean variable to indicate whether the graphic is a
representation of frequencies}

\item{main}{title for the plot}

\item{xlim}{limits for the x-axis}

\item{ylim}{limits for the y-axis}
}
\value{
a list with
            f   the frequency for each midpoint
            x   the midpoints of the nBins intervals
}
\description{
Generates histogram
}
\author{
Ram Ahluwalia \email{ram@wingedfootcapital.com} and Xavier Valls
\email{xaviervallspla@gmail.com}
}
\references{
\url{http://www.symmys.com}
See Meucci script pHist.m used for plotting
}

