\name{PlotCompositionEfficientFrontier}
\alias{PlotCompositionEfficientFrontier}
\title{Plots the efficient frontier}
\usage{
  PlotCompositionEfficientFrontier(Portfolios)
}
\arguments{
  \item{Portfolios}{: [matrix] (M x N) M portfolios of size
  N (weights)}
}
\description{
  Plot the efficient frontier, as described in A. Meucci,
  "Risk and Asset Allocation", Springer, 2005.
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  A. Meucci - "Exercises in Advanced Risk and Portfolio
  Management" \url{http://symmys.com/node/170}.

  See Meucci's script for
  "PlotCompositionEfficientFrontier.m"
}

