\name{PlotDistributions}
\alias{PlotDistributions}
\title{plot dists}
\usage{
  PlotDistributions(X, p, Mu, Sigma, p_, Mu_, Sigma_)
}
\arguments{
  \item{X}{}

  \item{p}{}

  \item{Mu}{}

  \item{Sigma}{}

  \item{p_}{}

  \item{Mu_}{}

  \item{Sigma_}{}
}
\description{
  plot dists
}
\author{
  Ram Ahluwalia \email{ram@wingedfootcapital.com}
}

