\name{PlotVolVsCompositionEfficientFrontier}
\alias{PlotVolVsCompositionEfficientFrontier}
\title{Plot the efficient frontier in the plane of portfolio weights versus standard deviation,
as described in  A. Meucci, "Risk and Asset Allocation", Springer, 2005.}
\usage{
  PlotVolVsCompositionEfficientFrontier(Portfolios, vol)
}
\arguments{
  \item{Portfolios:}{[matrix] (M x N) of portfolios
  weights}

  \item{vol}{: [vector] (M x 1) of volatilities}
}
\description{
  Plot the efficient frontier in the plane of portfolio
  weights versus standard deviation, as described in A.
  Meucci, "Risk and Asset Allocation", Springer, 2005.
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  \url{http://symmys.com/node/170} See Meucci's script for
  "PlotVolVsCompositionEfficientFrontier.m"
}

