% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudodata}
\alias{pseudodata}
\title{Pseudo data: Gauss-Hermite grid case study input data}
\description{
Pseudo data: Gauss-Hermite grid case study input data
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, D. Ardia, S. Keel - "Fully Flexible Extreme Views"
}
\keyword{data}

