% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuadraticMatVC.R
\name{QuadraticMat_Vc}
\alias{QuadraticMat_Vc}
\title{Computes the matrix q_t of the problem to solve when using
CALCULUS of VARIATION}
\usage{
QuadraticMat_Vc(lambda, gamma, eta, sig2, c2, tau_, n_, i_invest = NULL)
}
\arguments{
\item{lambda}{[scalar] discounting parameter}

\item{gamma}{[scalar] risk aversion parameter}

\item{eta}{[scalar] overall weight of the market impact of transactions}

\item{sig2}{[n_*t_ x n_*t_] covariance matrix of the process of the risk
drivers}

\item{c2}{[k_ x k_] market impact matrix}

\item{tau_}{[scalar] effective number of future time steps considered}

\item{n_}{[scalar] number of risk drivers}

\item{i_invest}{[k_ x 1] labels of the investible risk drivers}
}
\value{
q_t   [k_*tau_ x k_*tau_] matrix computed
}
\description{
This function computes the matrix q_t of the problem to solve
when using CALCULUS of VARIATION:
argmin_b (b' q_t b - b'l_t)
}
\note{
t_ = number of monitoring times at which sig2 is computed
k_ = number of investible risk drivers
}
\author{
Xavier Valls \email{xavievallspla@gmail.com}
}
\references{
A. Meucci - "Dynamic Portfolio Management with Views at Multiple Horizons"
\url{http://symmys.com/node/831}. See Meucci script for "MVOU_Prior.m"
}

