\docType{data}
\name{returnsDistribution}
\alias{returnsDistribution}
\title{Panel X of joint returns realizations and vector p of respective probabilities}
\description{
  Panel X of joint returns realizations and vector p of
  respective probabilities
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, "Fully Flexible Views: Theory and Practice",
  The Risk Magazine, October 2008, p 100-106.
  \url{http://symmys.com/node/158}
}
\keyword{data}
\keyword{datasets}

