\docType{data}
\name{sectorsSnP500}
\alias{sectorsSnP500}
\title{data from the sectors in the S&P 500}
\description{
  data from the sectors in the S&P 500
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  Attilio Meucci, 2011, "Robust Bayesian Allocation".
  \url{http://symmys.com/node/102}
}
\keyword{data}
\keyword{datasets}

