\docType{data}
\name{sectorsTS}
\alias{sectorsTS}
\title{stock returns by sectors}
\description{
  stock returns by sectors
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

