\docType{data}
\name{securitiesIndustryClassification}
\alias{securitiesIndustryClassification}
\title{Stock Indices}
\description{
  Stock Indices
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

