\name{StackedBarChart}
\alias{StackedBarChart}
\title{Generate a Stacked Bar Chart based on the frontier weights matrix}
\usage{
  StackedBarChart(weightsMatrix)
}
\arguments{
  \item{a}{matrix of weights where rows are efficient
  portfolios summing to one, and columns are assets}

  \item{a}{string indicating the title of the chart}
}
\description{
  Generate a Stacked Bar Chart based on the frontier
  weights matrix
}

