\name{std}
\alias{std}
\title{Calculates the population standard deviation}
\usage{
  std(x)
}
\arguments{
  \item{x}{a generic numeric vector}
}
\value{
  std a numeric with the population standard deviaiton of
  the generic numeric
}
\description{
  Calculates the population standard deviation dividing by
  'n' instead of 'n-1' equivalent to Matlab
}
\author{
  Ram Ahluwalia \email{rahluwalia@gmail.com}
}

