% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StockSeries}
\alias{StockSeries}
\title{Stock Series}
\description{
Stock Series
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, Exercises in Advanced Risk and Portfolio Management. \url{http://symmys.com/node/170}
}
\keyword{data}

