% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Torsion.R
\name{Torsion}
\alias{Torsion}
\title{Computes the Principal Components torsion and the Minimum Torsion}
\usage{
Torsion(Sigma, model, method = NULL, max_niter = 10000)
}
\arguments{
\item{Sigma}{[matrix] (n_ x n_) covariance matrix}

\item{model}{[string] choose between 'pca' and 'minimum-torsion'
model}

\item{method}{[string] choose between 'approximate' and 'exact'
method for 'minimum-torsion' model}

\item{max_niter}{[scalar] choose number of iterations of numerical
                               algorithm}
}
\value{
t             [matrix] (n_ x n_) torsion matrix
}
\description{
Computes the Principal Components torsion and the Minimum
Torsion for diversification analysis, as described in A. Meucci,
A. Santangelo, R. Deguest - "Measuring Portfolio Diversification Based on
Optimized Uncorrelated Factors"
}
\author{
Xavier Valls \email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, A. Santangelo, R. Deguest - "Measuring Portfolio Diversification
Based on Optimized Uncorrelated Factors" \url{http://symmys.com/node/599}

See Meucci's script "torsion.m"
}

