\name{Tweak}
\alias{Tweak}
\title{tweak a matrix}
\usage{
  Tweak(A, b, g)
}
\arguments{
  \item{A}{matrix A consisting of inequality constraints (
  Ax <= b )}

  \item{b}{matrix b consisting of inequality constraint
  vector b ( Ax <= b )}

  \item{g}{TODO: TBD}
}
\value{
  db
}
\description{
  tweak a matrix
}
\author{
  Ram Ahluwalia \email{ram@wingedfootcapital.com}
}

