\docType{data}
\name{UsSwapRates}
\alias{UsSwapRates}
\title{US Swap Rates}
\description{
  US Swap Rates
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

