\name{ViewCurveSlope}
\alias{ViewCurveSlope}
\title{Process views for the expectations and binding constraints as it appears in A. Meucci,
 "Fully Flexible Views: Theory and Practice", The Risk Magazine, October 2008,
 p 100-106}
\usage{
  ViewCurveSlope(X, p)
}
\arguments{
  \item{X}{: panel of joint factors realizations}

  \item{p}{: vector of probabilities}
}
\value{
  p_ : vector of posterior probabilities
}
\description{
  Process views for the expectations and binding
  constraints as it appears in A. Meucci, "Fully Flexible
  Views: Theory and Practice", The Risk Magazine, October
  2008, p 100-106
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  A. Meucci, "Fully Flexible Views: Theory and Practice"
  \url{http://www.symmys.com/node/158} See Meucci script
  for "ButterflyTrading/ViewCurveSlope.m"
}

