\name{ViewRanking}
\alias{ViewRanking}
\title{view the rankings}
\usage{
  ViewRanking(X, p, Lower, Upper)
}
\arguments{
  \item{X}{a vector containing returns for all the asset
  classes}

  \item{p}{a vector containing the prior probability
  values}

  \item{Lower}{a vector of indexes indicating which column
  is lower than the corresponding column number in Upper}

  \item{Upper}{a vector of indexes indicating which column
  is lower than the corresponding column number in Upper}
}
\description{
  view the rankings
}

