\docType{data}
\name{butterfliesAnalytics}
\alias{butterfliesAnalytics}
\title{list of securities with analytics computed.}
\description{
  list of securities with analytics computed.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, "Fully Flexible Views: Theory and Practice",
  The Risk Magazine, October 2008, p 100-106.
  \url{http://symmys.com/node/158}
}
\keyword{data}
\keyword{datasets}

