\docType{data}
\name{derivatives}
\alias{derivatives}
\title{implied vol for options on SPX}
\description{
  implied vol for options on SPX for different time to
  maturity and moneyness.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

