% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factorsDistribution}
\alias{factorsDistribution}
\title{Panel X of joint factors realizations and vector p of respective probabilities for factors}
\description{
Panel X of joint factors realizations and vector p of respective probabilities for factors
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, "Fully Flexible Views: Theory and Practice", The Risk Magazine,
October 2008, p 100-106. \url{http://symmys.com/node/158}
}
\keyword{data}

