% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fILMR}
\alias{fILMR}
\title{Fully Integrated Liquidity and Market Risk Model dataset.}
\description{
Data for the Fully Integrated Liquidity and Market Risk Model paper.
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
A. Meucci, "A Fully Integrated Liquidity and Market Risk Model", Financial Analyst Journal, 68, 6, 35-47 (2012) \url{http://www.symmys.com/node/350}
}
\keyword{data}

