\name{garch2f8}
\alias{garch2f8}
\title{Off-diagonal parameter estimation in bivariate GARCH(1,1) when diagonal parameters are given.}
\usage{
  garch2f8(y, c1, a1, b1, y1, h1, c2, a2, b2, y2, h2, df)
}
\arguments{
  \item{y}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}

  \item{c1}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix C}

  \item{a1}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix A}

  \item{b1}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix B}

  \item{y1}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}

  \item{h1}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}

  \item{c2}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix C}

  \item{a2}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix A}

  \item{b2}{: [scalar] diagonal parameter of the GARCH(1,1)
  process taken from matrix B}

  \item{y2}{: [vector] (T x 1) data generated by a
  GARCH(1,1) process}

  \item{h2}{: [vector] (T x 1) generated by a GARCH(1,1)
  process}

  \item{df}{: [scalar] degree of freedom for the
  t-distribution; the default value is 500 to make it,
  basically, normal}
}
\value{
  q : [vector] (4 x 1) parameters of the GARCH(1,1) process

  qerr : [vector] (4 x 1) standard error of parameter
  estimates

  hf : [scalar] current conditional heteroskedasticity
  estimate

  hferr : [scalar] standard error on hf
}
\description{
  Off-diagonal parameter estimation in bivariate GARCH(1,1)
  when diagonal parameters are given.
}
\note{
  MATLAB's code initially written by Olivier Ledoit,
  4/28/1997

  Uses a conditional t-distribution with fixed degrees of
  freedom

  Steepest Ascent on boundary, Hessian off boundary, no
  grid search
}
\author{
  Xavier Valls \email{flamejat@gmail.com}
}
\references{
  A. Meucci - "Exercises in Advanced Risk and Portfolio
  Management" \url{http://symmys.com/node/170}.

  See Meucci's script for "FitMultivariateGarch.m"
}

