\docType{data}
\name{linearModel}
\alias{linearModel}
\title{parameters for the explicit factors / implicit loadings linear model}
\description{
  parameters for the explicit factors / implicit loadings
  linear model.
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

