\name{PlotDistributions}
\alias{PlotDistributions}
\title{Plot numerical and analytical prior and posterior distributions}
\usage{
  PlotDistributions(X, p, Mu, Sigma, p_, Mu_, Sigma_)
}
\arguments{
  \item{X}{a vector containing the dataset}

  \item{p}{a vector cotaining the prior probability values}

  \item{Mu}{a vector containing the prior means}

  \item{Sigma}{a vector containing the prior standard
  deviations}

  \item{p_}{a vector containing the posterior probability
  values}

  \item{Mu_}{a vector containing the posterior means}

  \item{Sigma_}{a vector containing the posterior standard
  deviations}
}
\description{
  Plot numerical and analytical prior and posterior
  distributions
}
\author{
  Ram Ahluwalia \email{ram@wingedfootcapital.com}
}

