\docType{data}
\name{swap2y4y}
\alias{swap2y4y}
\title{Swaps for 2y and 4y}
\description{
  Swaps for 2y and 4y
}
\author{
  Xavier Valls\email{flamejat@gmail.com}
}
\references{
  A. Meucci, Exercises in Advanced Risk and Portfolio
  Management. \url{http://symmys.com/node/170}
}
\keyword{data}
\keyword{datasets}

